déplace du code et retire du code inutile
